Delegate LDAP attributes to LdapUser
Creates a new LdapUser class that wraps the result from PsuLdapService
so we can interact with its attributes more easily. The User class then
delegates the methods it needs to this class enabling us to use the
build-in ActiveModel::Serializers for json representations of user data.
This cleans up the UserController by ceding all of that logic to the
User class.

In the process, there's a lot of refactoring and some additional gems
like Faker and rspec-its to aid in the testing.